requirements.txt
fastapi
uvicorn
pydantic
pillow
google-generativeai
requests
pdf2image
pytesseract
